Had previously meant to upcast to be sure that the correct pointer was stored in data.
Because window_t and therefore window_base didn't have to be the first class inherited by the window, we have to be sure that the pointer passed to CreateWindowEx is something that we can work with. Therefore, the pointer passed to CreateWindow is upcast to window_base. This also allows us to use the instance in window_base too.

When we get the instance, first we cast to window_base and then downcast to our derived type.